use selector for global alerts
